Some servers may not have net tools like netstat installed. This is the case with many docker images for instance. This commit removes the use of netstat to determing if the service is listening on its designated port. Ultimately this check is unnecessary given that immediately afterwards the http endpoint is checked which listens on the same port and accomplishes the same check.
This commit also ensures that the server recipe always performs this check instead of only performing it when the service resource is started. The couchbase installer may start the service causing the service resource to skip and then the check is never performed resulting in a failure on the first converge.